Move helper methods from conftest into new module _conftest_helpers.py
