+++ pyaerocom/conftest.py
@@ -8,91 +8,24 @@ Created on Tue Feb 11 15:57:09 2020
 import pytest
 
 from pathlib import Path
-import os
-import requests
-import tarfile
 
 from traceback import format_exc
+
 from pyaerocom import const
+import pyaerocom._conftest_helpers as cth                          
 from pyaerocom.griddeddata import GriddedData
 from pyaerocom.io import ReadAasEtal
 from pyaerocom.io import ReadAeronetSunV3, ReadAeronetSdaV3, ReadAeronetInvV3
 from pyaerocom.io import ReadEbas
 from pyaerocom.test.synthetic_data import DataAccess
 
-def _download_test_data(basedir=None):
-    print("TEMP OUTPUT: DOWNLOADING TESTDATA")
-    #raise Exception('Temporarily disabled...')
-    if basedir is None:
-        basedir = const.OUTPUTDIR
-   
-    download_loc = Path(basedir).joinpath('{}.tar.gz'.format(const._testdatadirname))
-    
-    try:
-        r = requests.get(_URL_TESTDATA)
-        with open(download_loc, 'wb') as f:
-            f.write(r.content) 
-            
-        with tarfile.open(download_loc, 'r:gz') as tar:
-            tar.extractall(const.OUTPUTDIR)
-            tar.close()
-    except Exception:
-        const.print_log.warning('Failed to download testdata. Traceback:\n{}'
-                                .format(format_exc()))
-        return False
-    finally:
-        if download_loc.exists():
-            os.remove(download_loc)
-    return True
-
-def _check_access_testdata(basedir, test_paths):
-    if isinstance(basedir, str):
-        basedir = Path(basedir)
-    elif not isinstance(basedir, Path):
-        raise ValueError('Invalid input for basedir ({}), need str or Path'
-                         .format(type(basedir)))
-    if not basedir.exists():
-        return False
-    
-    if not isinstance(test_paths, dict):
-        raise ValueError('Invalid input for test_paths, need dict')
-    
-    for data_id, data_dir in test_paths.items():
-        if not basedir.joinpath(data_dir).exists():
-            return False
-    return True
-                
-def check_access_testdata(basedir, test_paths):
-    if not _check_access_testdata(basedir, test_paths):
-        try:
-            if _download_test_data(const.OUTPUTDIR):
-                if _check_access_testdata(basedir, test_paths):
-                    return True
-        except Exception:
-            pass
-        return False
-    return True
-
-def _init_testdata(const):
-    for name, relpath in ADD_PATHS.items():
-        ddir = str(TESTDATADIR.joinpath(relpath))
-        if name in _UNGRIDDED_READERS:
-            reader = _UNGRIDDED_READERS[name]
-            
-            const.add_ungridded_obs(name, ddir, 
-                                    reader=reader,
-                                    check_read=True)
-            
-        else:
-            const.add_data_search_dir(ddir)
-            
 INIT_TESTDATA = True                            
 TEST_RTOL = 1e-4
 
 DATA_ACCESS = DataAccess()
 
 # thats were the testdata can be downloaded from
-_URL_TESTDATA = 'https://pyaerocom.met.no/pyaerocom-suppl/testdata-minimal.tar.gz'
+URL_TESTDATA = 'https://pyaerocom.met.no/pyaerocom-suppl/testdata-minimal.tar.gz'
 
 # Testdata directory
 TESTDATADIR = Path(const._TESTDATADIR)
@@ -135,11 +68,12 @@ NASA_AMES_FILEPATHS = {
 # checks if testdata-minimal is available and if not, tries to download it 
 # automatically into ~/MyPyaerocom/testdata-minimal
 if INIT_TESTDATA:
-    TESTDATA_AVAIL = check_access_testdata(TESTDATADIR, TEST_PATHS)
+    TESTDATA_AVAIL = cth.check_access_testdata(TESTDATADIR, TEST_PATHS)
     
     if TESTDATA_AVAIL:
         try:
-            _init_testdata(const)
+            cth._init_testdata(const, ADD_PATHS, TESTDATADIR, 
+                               _UNGRIDDED_READERS)
         except Exception:
             raise ValueError('FATAL: Failed to initiate testdata. Traceback:\n'
                              .format(format_exc()))
